+++ src/lib/data/constant.ts
@@ -33,14 +33,18 @@ export const typeUrlDict = {
 export const DEFAULT_RPC_ERROR = "Invalid format, or Something went wrong";
 
 export const DEFAULT_TX_FILTERS = {
+  isSend: false,
+  isIbc: false,
   isExecute: false,
   isInstantiate: false,
   isUpload: false,
-  isIbc: false,
-  isSend: false,
   isMigrate: false,
   isUpdateAdmin: false,
   isClearAdmin: false,
+  isPublish: false,
+  isEntryExecute: false,
+  isUpgrade: false,
+  isScript: false,
 };
 
 export const UPPERBOUND_COUNT = 10000;